refactor: organise search stuff away from FHIR
