+++ packages/gateway/src/features/search/root-resolvers.ts
@@ -10,13 +10,8 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import { ApiResponse } from '@elastic/elasticsearch'
-import {
-  getMetrics,
-  postAdvancedSearch,
-  postMetrics
-} from '@gateway/features/fhir/utils'
+import { getMetrics, postMetrics } from '@gateway/features/fhir/utils'
 import { markRecordAsDownloadedOrAssigned } from '@gateway/features/registration/root-resolvers'
-import { ISearchCriteria } from '@gateway/features/search/type-resolvers'
 import {
   getSystem,
   getTokenPayload,
@@ -25,6 +20,7 @@ import {
 } from '@gateway/features/user/utils'
 import { GQLResolver } from '@gateway/graphql/schema'
 import { Options } from '@hapi/boom'
+import { ISearchCriteria, postAdvancedSearch } from './utils'
 
 // Complete definition of the Search response
 interface IShardsResponse {
@@ -118,11 +114,7 @@ export const resolvers: GQLResolver = {
         const searchResult: ApiResponse<ISearchResponse<any>> =
           await postAdvancedSearch(authHeader, searchCriteria)
 
-        if (
-          searchResult &&
-          searchResult.statusCode &&
-          searchResult.statusCode >= 400
-        ) {
+        if ((searchResult?.statusCode ?? 0) >= 400) {
           const errMsg = searchResult as Options<string>
           return await Promise.reject(new Error(errMsg.message))
         }
@@ -136,16 +128,8 @@ export const resolvers: GQLResolver = {
         }
 
         return {
-          totalItems:
-            (searchResult &&
-              searchResult.body.hits &&
-              searchResult.body.hits.total.value) ||
-            0,
-          results:
-            (searchResult &&
-              searchResult.body.hits &&
-              searchResult.body.hits.hits) ||
-            []
+          totalItems: searchResult.body.hits.total.value ?? 0,
+          results: searchResult.body.hits.hits ?? []
         }
       } else {
         const hasAtLeastOneParam = Object.values(advancedSearchParameters).some(
@@ -174,8 +158,8 @@ export const resolvers: GQLResolver = {
         const searchResult: ApiResponse<ISearchResponse<any>> =
           await postAdvancedSearch(authHeader, searchCriteria)
         return {
-          totalItems: searchResult?.body?.hits?.total?.value || 0,
-          results: searchResult?.body?.hits?.hits || []
+          totalItems: searchResult?.body?.hits?.total?.value ?? 0,
+          results: searchResult?.body?.hits?.hits ?? []
         }
       }
     },
