+++ packages/gateway/src/features/search/type-resolvers.ts
@@ -14,10 +14,8 @@ import {
   NATIVE_LANGUAGE,
   USER_MANAGEMENT_URL
 } from '@gateway/constants'
-import { resolve } from 'url'
 import fetch from 'node-fetch'
 import {
-  GQLAdvancedSearchParametersInput,
   GQLOperationHistorySearchSet,
   GQLResolver
 } from '@gateway/graphql/schema'
@@ -35,14 +33,6 @@ interface ISearchEventDataTemplate {
 }
 interface ISearchDataTemplate {
   [key: string]: any
-}
-export interface ISearchCriteria {
-  parameters: GQLAdvancedSearchParametersInput
-  sort?: string
-  sortColumn?: string
-  size?: number
-  from?: number
-  createdBy?: string
 }
 
 interface IAssignment {
@@ -443,7 +433,7 @@ export const searchTypeResolvers: GQLResolver = {
   AssignmentData: {
     async avatarURL(assignmentData: IAssignment, _, { headers: authHeader }) {
       const response = await fetch(
-        resolve(USER_MANAGEMENT_URL, `users/${assignmentData.userId}/avatar`),
+        new URL(`users/${assignmentData.userId}/avatar`, USER_MANAGEMENT_URL),
         {
           method: 'GET',
           headers: {
