+++ packages/gateway/src/features/search/utils.ts
@@ -1,0 +1,45 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+import { IAuthHeader } from '@gateway/common-types'
+import { SEARCH_URL } from '@gateway/constants'
+import { GQLAdvancedSearchParametersInput } from '@gateway/graphql/schema'
+import fetch from 'node-fetch'
+
+export interface ISearchCriteria {
+  parameters: GQLAdvancedSearchParametersInput
+  sort?: string
+  sortColumn?: string
+  size?: number
+  from?: number
+  createdBy?: string
+}
+
+export const postAdvancedSearch = async (
+  authHeader: IAuthHeader,
+  criteria: ISearchCriteria
+) => {
+  try {
+    const response = await fetch(`${SEARCH_URL}advancedRecordSearch`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        ...authHeader
+      },
+      body: JSON.stringify(criteria)
+    })
+    return await response.json()
+  } catch (error) {
+    return await Promise.reject(
+      new Error(`Search request failed: ${error.message}`)
+    )
+  }
+}
